更新 galileo.yaml 配置，重构 Titan 和 DFlow 引擎支持
- 在 galileo.yaml 中新增 Titan 和 DFlow 引擎的配置，优化多 IP 订阅逻辑，提升交易性能。
- 调整引擎参数，确保每个 IP 仅维持两条固定订阅，增强系统的可扩展性。
- 更新相关文档，详细描述新的引擎配置和使用方法，提升用户体验。
- 在代码中新增对地址查找表的支持，增强交易解码功能，提供更丰富的交易信息输出。

+++ src/api/ultra/order.rs
@@ -49,7 +49,7 @@ pub struct SwapInfo {
 #[serde(rename_all = "camelCase")]
 pub struct RoutePlanStep {
     pub swap_info: SwapInfo,
-    pub percent: u8,
+    pub percent: Decimal,
     #[serde(skip_serializing_if = "Option::is_none")]
     pub bps: Option<u16>,
 }
